demo: move mtest opponent to separate file
